fix: generate correct number of random points

+++ src/settings.cpp
@@ -154,14 +154,14 @@ void Settings::showSettings(bool *showSettings, float toolbarHeight, float *numb
             GuiLabel((Rectangle){window_position.x + 20 + scroll.x, window_position.y + 50 + scroll.y, 250, 25},
                      "Number of Points");
             GuiSlider((Rectangle){window_position.x + 20 + scroll.x, window_position.y + 100 + scroll.y, 300, 25}, NULL,
-                      TextFormat("%1.0f", *numberOfPoints), numberOfPoints, 0.0f, 100.0f);
+                      TextFormat("%d", (int)(*numberOfPoints)), numberOfPoints, 0.0f, 100.0f);
 
             if (GuiButton((Rectangle){window_position.x + 20 + scroll.x, window_position.y + 150 + scroll.y, 150, 25},
                           "Generate"))
             {
                 dataPoints.clear();
                 int i = 0;
-                while (i < *numberOfPoints)
+                while (i < (int)(*numberOfPoints))
                 {
                     int x = rand() % GetScreenWidth(), y = rand() % GetScreenHeight();
                     if (y > toolbarHeight + 10 && y < GetScreenWidth() - 10 && x > 10 && x < GetScreenWidth() - 10)
